requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
chromadb==0.5.5
groq==0.11.0
huggingface_hub==0.24.7
httpx==0.27.0

